requirements.txt
# Google ADK and Vertex AI
google-genai>=0.3.0
google-cloud-aiplatform>=1.70.0

# FastAPI and Web Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# RAG Components
faiss-cpu>=1.8.0
pypdf>=5.0.0
langchain>=0.3.0
langchain-text-splitters>=0.3.0

# Web Search
tavily-python>=0.5.0

# Logging and Observability
python-json-logger>=2.0.7
google-cloud-logging>=3.11.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=24.1.0
requests>=2.32.0
